Pass `EnableProfileLoading` and `InitialWorkingDirectory` as `initializationOptions` (#3986)
To fix stdio clients, we needed to fix our startup logic so that the
PowerShell host was actually started on initialization, not in the
`onDidChangeConfiguration` handler. The previous way it was done only
worked because it relied on a quirk of the VS Code client to send that
request immediately, which other clients don’t necessarily do. In order
to startup sooner though, we need any configuration that’s applicable on
startup to be passed across the wire with the `initialize` request.