Don't hide extension terminal entirely (#4096)
When `showOnStartup` is false, we currently completely hide the terminal
from the user via an API that essentially backgrounds the terminal. This
was in response to a user request, but has caused more problems than
it's worth. This PR reverts the behavior so that it simply won't show
the terminal pane on startup, but opening it reveals the connected
extension terminal.