Improve language client library close action message
The updated language client library now displays an error when the
connection to the server terminates. Thing is, we already have our own
restart experience where we inform the user that the terminal should be
restarted, and give them a Yes/No option to do so. Unfortunately, we
can't suppress this error message, and I don't think that we want to
auto-restart the terminal, so I've at least made the message a little
more clear.