Add `integratedConsole.startInBackground` to completely hide the terminal (#4152)
This makes hiding the Extension Terminal from the set of terminals in VS
Code an option, with a big warning that most users are probably just
looking for `showOnStartup`, and that if enabled, to access the terminal
you must manually open it some other way (such as with `Show Extension
Terminal`). Some users require this as part of their workflow.