Refactor the debug launch configuration resolvers (#4172)
Some much needed TLC, especially now that VS Code with resolve its
predefined variables used in configurations.

Do not set `cwd` for debug launch configurations

Instead, respect it if it exists, and otherwise consistently do not
change it. Debugging will run in the session's current `cwd`, or if and
only if the user set `cwd` on the launch config will it change.