Fix session and log directories when user folder is on UNC share
When we moved to using `globalStorageUri` and passing `vscode.Uri`s
around, we did not test UNC shares. It turns out that while
`vscode.Uri.fsPath`'s documentation says that it "does not look at
scheme," it actually does look at the scheme and essentially requires it
to be `file`. The scheme of the `globalStorageUri` however is
`vscode-userdata`, so the `fsPath` fails to then handle the UNC share
path. We are now replacing the scheme with `file` to get around this bug.