Skip auto-update on unsupported Windows architectures (#4298)
While PowerShell now works on ARM64 Windows, it does not yet have an MSI
installer, so for now instead of erroneously attempting to install the
x64 version, we just recommend that the user update (like we do for
Linux). Note that the string is now capitalized due to a server-side
change that simply returns .NET's
`System.Runtime.InteropServices.Architecture` enum as a string.