Restore original settings after disabling ISE mode
We were previously setting them to `undefined` which mostly worked,
unless the user had customized the setting. This was most apparent and
commonly done with `colorTheme`. Now we save when we enable, and restore
to that value. Note that will only work for the original session, but it
does allow the user to test ISE mode out without repercussions.

We could probably save the settings into the global storage context, but
that might get hairy!