Remove `preview` label from marketplace
We were using this as an easy way to determine if the version is a pre-release.
However, the marketplace always looks a the latest version meaning it was
placing a "preview" label and our preview icon on the webpage.

So we had to write a different way to determine if the extension is a pre-release.