Respect file path casing in extension commands
Per https://nodejs.org/en/docs/guides/working-with-different-filesystems
we should really just keep the file path we're given, no "normalizing"
to be done especially since it is wrong to infer from the OS if the
filesystem is case-sensitive or not. Gone are the days where Windows and
macOS could be presumed case-insensitive, they now both support and are
often found with with case-sensitive filesystems. Plus, there was really
no reason to be doing this in the first place. Tested interactively,
this works just fine since the VS Code APIs we're using handle
case-insensitivity for us.

Resolves #2960.