Report errors from extension API commands
The response is now `Completed` or `Failed` and returned appropriately.
`NewFile` now takes (optional) content and fills the new file with it
(which is all the VS Code API can do). `OpenFile` doesn't throw a
cryptic error in the terminal if the file isn't found. The status bar
messages are now properly disposed.