Update stale action logic
There was a race condition because this acts on an hourly poll, so if a
user replied to an issue that "Needs: Author Feedback" within an
hour...it would still get marked stale. Instead, we're doing away with
the "Stale" label and simplifying our workflow to just adding either a
resolved label (which will then have a "Needs: Fix Verification" added
as the stale label) or just "Needs: Author Feedback" which is the stale
label. For resolved issues, if it remains stale for a week, it will
close. If it's updated it'll have "Needs: Maintainer Attention" applied
and then exempted from closing so we can check. For issues we've marked
as "Needs: Author Feedback" since this is the stale label, it won't
automatically add anything, which means instead it's immediately stale
from the moment we add the label (eliminating the race condition). So it
should shuffle the labels for us correctly and close after a week.