Fix up debugger configuration resolvers
Ok so we removed the notion of IDs (which are integers) being strings
because that broke serialization with recent updates. But on inspection
we realized the whole point was to be able to say "current" and have the
attach configuration attach to the currently running Extension Terminal.
Except that this was an unsupported, half-baked scenario (sure, it
attached, but it couldn't actually be debugged properly). So now that
throws warnings in both the client and server. The defaults for
`processId` and `runspaceId` where changed to `null` (again, now they're
just integers) which means to query for them. The supplied but
half-baked configuration was removed.