Refactor `LanguageClientConsumer` to wait properly
This makes a lot more sense and fixes a long-standing bug.  Also cleaned up a
lot while at it.

1. `SessionManager` is created and the features are injected into the session
manager with the intention of them being able to "register" their language
client event handlers in the period between when the client is created and when
it is started.

2. For each feature, if it gets to a point where it needs the language client,
we want them to wait on a promise for the "started" client to be provided from
the session manager before continuing.

3. Once it is started, the session manager passes the client to
`LanguageClientConsumer` which resolves the promise for the features so they can
continue.