Disable Azure Pipelines CI for release branch
It is completely superfluous because the release pipeline runs the exact same CI
template. Moreover, because it's a different Azure organization, the NPM
authentication won't work without setting up and managing a service connection.

Run the GitHub Actions CI with the Release configuration.

Also run the NPM install scripts because sometimes it will break without them.