Pass the path to `shellIntegration.ps1` directly to the server
The server can't use `code --locate-shell-integration-path pwsh`
because it doesn't know if it's `code` or `code-insiders` etc.
Even when given the direct path to the client's host process, it can't
reliably use that API either. Mostly this is due to Windows using
`Code.cmd` in the background (which is not `Code.exe`, the host process)
but it's also slower anyway as it requires another launch of Node.js.

We'll have to rely on the path to the script not changing relative to
`vscode.env.appRoot`, but at least that part is a public API.