Fix OneBranch pipeline
I mindlessly copied the old pipeline's logic and had it building from
the Git repository. However, weirdly the NuGet authentication stopped
working, and it was a better fix to test against the same artifact
that we're going to release.

We also have to release from the `main` branch due to _reasons_,
and this was actually fortunate as I noticed I'd made a mistake with the
VSIX artifact names (they're not prefixed with `vscode-`).