Reorganize build and packages to allow for omitting dev tools
New requirements in our release pipeline require us to omit the installation
of quite a few of our (optional) dev tools, as their post-install tasks
reach out to the public NPM registry and so are blocked. By moving everything
except `esbuild` and `vsce` to `optionalDependencies` (and reorganzing the
build script) we can run the linter and tests outside of the secure container.

Since OneBranch and GitHub Actions are now using PowerShell 7.4, the module
setup script had to be updated. Because OneBranch now runs without releasing,
the downloaded drop had to change (skipping a zip and extraction).