v2024.5.1-preview: Drop support for PowerShell <7.4 and logging overhaul
PowerShell 7.2 LTS and 7.3 are now past end-of-support and are now unsupported.
This is an incompatible API change so we're bumping the major version
of PowerShell Editor Services.
Please update to PowerShell 7.4 LTS going forward.

This release contains a logging overhaul which purposely removes our
dependency on Serilog and should lead to improved stability with
PowerShell 5.1 (by avoiding a major GAC assembly conflict).