Use correct build of PowerShellEditorServices in OneBranch
Since OneBranch is still using PowerShell 7.3, the PowerShellEditorServices
is only partially succeeding as its "assert release configuration" step
fails (I've verified it manually) with a continueOnError flag set to true.
So the correct build is from a pipeline that's only partially succeeded,
and the default setting of the download tasks silently skips those.
Now we explicitly allow them and print out the version so this is easier.