Improve warning when additional PowerShell isn't found
Since we added logic which searches for possible intended permutations
of the given additional PowerShell path, we needed to make the warning
show only if none of the permutations were found. This was accomplished
by suppressing it in the first iterator and then yielding it again after
the permutations were exhausted with the warning unsuppressed.