Add `temporaryConsoleWindowActionOnDebugEnd` option (#5255)
Adds a new option `temporaryConsoleWindowActionOnDebugEnd` to both the
attach and launch configurations which can be used to specify what
happens with the temporary integrated console when a debug session ends.
The option can be set to `keep`, current behaviour and default, that
will keep the active terminal as the temporary console. It can be set to
`close` which closes the terminal and removes it from the selection
pane or `hide` which keeps the terminal window alive but changes the
active terminal to the previous one before the debug session started.